reorganize file structure, add dev tools, and renamed extension
