+++ src/DevTool/DevtoolsApp.tsx
@@ -1,10 +1,12 @@
 import React from 'react'
+import PopUpApp from '../PopUp/PopUpApp'
+import "../App.css"
+
 
 function DevToolsApp() {
   return (
     <div className="p-4">
-      <h1 className="text-2xl mb-4">DevTools Panel</h1>
-      <p>This is a placeholder for the DevTools panel.</p>
+      <PopUpApp />
     </div>
   )
 }
