+++ src/PopUp/PopUpApp.jsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom'
 
-import './App.css'
+import '../App.css'
 
 function PopUpApp() {
   return (
