+++ src/PopUp/routes/js-files.jsx
@@ -1,4 +1,4 @@
-import { NavBar } from '../components/navbar'
+import { NavBar } from '../../components/navbar'
 
 export function JSFiles(){
     return(
