+++ src/PopUp/routes/urls.jsx
@@ -1,4 +1,4 @@
-import { NavBar } from '../components/navbar'
+import { NavBar } from '../../components/navbar'
 export function URLS(){
     return(
         <div>
